# Fix go 1.14's checkptr validation